pyproject.toml
[tool.poetry]
name = "stickit-api-server"
version = "0.0.1"
description = ""
authors = ["CHOI HEE JAE <gmlwo530@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.8"
appdirs = "1.4.4"
click = "7.1.2"
dnspython = "2.1.0"
email-validator = "1.1.2"
fastapi = "0.63.0"
h11 = "0.12.0"
mccabe = "0.6.1"
mypy-extensions = "0.4.3"
pathspec = "0.8.1"
pycodestyle = "2.6.0"
pydantic = "1.8"
pyflakes = "2.2.0"
python-multipart = "0.0.5"
regex = "2020.11.13"
six = "1.15.0"
starlette = "0.13.6"
toml = "0.10.2"
typed-ast = "1.4.2"
typing-extensions = "3.7.4.3"
uvicorn = "0.13.4"
requests = "^2.25.1"
python-dotenv = "^0.17.0"
motor = "^2.4.0"
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyJWT = "^2.1.0"
httpx = "^0.18.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
awscli = "^1.19.17"
pytest = "^6.2.2"
pytest-asyncio = "^0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
